core-services/ai-fastapi/requirements.txt
fastapi
uvicorn[standard]
google-genai
httpx